migrate middleware handling
