+++ packages/next/server/next-server.ts
@@ -1,19 +1,33 @@
 import type { Params, Route } from './router'
-import type { CacheFs } from '../shared/lib/utils'
-import type { NextParsedUrlQuery, NextUrlWithParsedQuery } from './request-meta'
+import { CacheFs, execOnce } from '../shared/lib/utils'
+import {
+  addRequestMeta,
+  getRequestMeta,
+  NextParsedUrlQuery,
+  NextUrlWithParsedQuery,
+} from './request-meta'
+import type { MiddlewareManifest } from '../build/webpack/plugins/middleware-plugin'
 import type RenderResult from './render-result'
+import type { FetchEventResult } from './web/types'
+import type { ParsedNextUrl } from '../shared/lib/router/utils/parse-next-url'
 
 import fs from 'fs'
 import { join, relative } from 'path'
 import { IncomingMessage, ServerResponse } from 'http'
 
-import { PAGES_MANIFEST, BUILD_ID_FILE } from '../shared/lib/constants'
+import {
+  PAGES_MANIFEST,
+  BUILD_ID_FILE,
+  SERVER_DIRECTORY,
+  MIDDLEWARE_MANIFEST,
+} from '../shared/lib/constants'
 import { PagesManifest } from '../build/webpack/plugins/pages-manifest-plugin'
 import { recursiveReadDirSync } from './lib/recursive-readdir-sync'
 import { format as formatUrl, UrlWithParsedQuery } from 'url'
 import compression from 'next/dist/compiled/compression'
 import Proxy from 'next/dist/compiled/http-proxy'
 import { route } from './router'
+import { run } from './web/sandbox'
 
 import {
   BaseNextRequest,
@@ -27,6 +41,7 @@ import { ParsedUrlQuery } from 'querystring'
 import { apiResolver } from './api-utils'
 import { RenderOpts, renderToHTML } from './render'
 import { ParsedUrl } from '../shared/lib/router/utils/parse-url'
+import * as Log from '../build/output/log'
 
 import BaseServer, {
   FindComponentsResult,
@@ -36,8 +51,13 @@ import BaseServer, {
 import { getMiddlewareInfo, getPagePath, requireFontManifest } from './require'
 import { normalizePagePath } from './normalize-page-path'
 import { loadComponents } from './load-components'
-import isError from '../lib/is-error'
+import isError, { getProperError } from '../lib/is-error'
 import { FontManifest } from './font-utils'
+import { toNodeHeaders } from './web/utils'
+import { relativizeURL } from '../shared/lib/router/utils/relativize-url'
+import { parseNextUrl } from '../shared/lib/router/utils/parse-next-url'
+import { prepareDestination } from '../shared/lib/router/utils/prepare-destination'
+import { normalizeLocalePath } from '../shared/lib/i18n/normalize-locale-path'
 
 export * from './base-server'
 
@@ -568,5 +588,295 @@ export default class NextNodeServer extends BaseServer {
     serverless: boolean
   }) {
     return getMiddlewareInfo(params)
+  }
+
+  protected getMiddlewareManifest(): MiddlewareManifest | undefined {
+    if (!this.minimalMode) {
+      const middlewareManifestPath = join(
+        join(this.distDir, SERVER_DIRECTORY),
+        MIDDLEWARE_MANIFEST
+      )
+      return require(middlewareManifestPath)
+    }
+    return undefined
+  }
+
+  protected generateCatchAllMiddlewareRoute(): Route | undefined {
+    if (this.minimalMode) return undefined
+
+    return {
+      match: route('/:path*'),
+      type: 'route',
+      name: 'middleware catchall',
+      fn: async (req, res, _params, parsed) => {
+        if (!this.middleware?.length) {
+          return { finished: false }
+        }
+
+        const initUrl = getRequestMeta(req, '__NEXT_INIT_URL')!
+        const parsedUrl = parseNextUrl({
+          url: initUrl,
+          headers: req.headers,
+          nextConfig: {
+            basePath: this.nextConfig.basePath,
+            i18n: this.nextConfig.i18n,
+            trailingSlash: this.nextConfig.trailingSlash,
+          },
+        })
+
+        if (!this.middleware?.some((m) => m.match(parsedUrl.pathname))) {
+          return { finished: false }
+        }
+
+        let result: FetchEventResult | null = null
+
+        try {
+          result = await this.runMiddleware({
+            request: req,
+            response: res,
+            parsedUrl: parsedUrl,
+            parsed: parsed,
+          })
+        } catch (err) {
+          if (isError(err) && err.code === 'ENOENT') {
+            await this.render404(req, res, parsed)
+            return { finished: true }
+          }
+
+          const error = getProperError(err)
+          console.error(error)
+          res.statusCode = 500
+          this.renderError(error, req, res, parsed.pathname || '')
+          return { finished: true }
+        }
+
+        if (result === null) {
+          return { finished: true }
+        }
+
+        if (result.response.headers.has('x-middleware-rewrite')) {
+          const value = result.response.headers.get('x-middleware-rewrite')!
+          const rel = relativizeURL(value, initUrl)
+          result.response.headers.set('x-middleware-rewrite', rel)
+        }
+
+        if (result.response.headers.has('Location')) {
+          const value = result.response.headers.get('Location')!
+          const rel = relativizeURL(value, initUrl)
+          result.response.headers.set('Location', rel)
+        }
+
+        if (
+          !result.response.headers.has('x-middleware-rewrite') &&
+          !result.response.headers.has('x-middleware-next') &&
+          !result.response.headers.has('Location')
+        ) {
+          result.response.headers.set('x-middleware-refresh', '1')
+        }
+
+        result.response.headers.delete('x-middleware-next')
+
+        for (const [key, value] of Object.entries(
+          toNodeHeaders(result.response.headers)
+        )) {
+          if (key !== 'content-encoding' && value !== undefined) {
+            res.setHeader(key, value)
+          }
+        }
+
+        const preflight =
+          req.method === 'HEAD' && req.headers['x-middleware-preflight']
+
+        if (preflight) {
+          res.statusCode = 200
+          res.send()
+          return {
+            finished: true,
+          }
+        }
+
+        res.statusCode = result.response.status
+        res.statusMessage = result.response.statusText
+
+        const location = result.response.headers.get('Location')
+        if (location) {
+          res.statusCode = result.response.status
+          if (res.statusCode === 308) {
+            res.setHeader('Refresh', `0;url=${location}`)
+          }
+
+          res.body(location).send()
+          return {
+            finished: true,
+          }
+        }
+
+        if (result.response.headers.has('x-middleware-rewrite')) {
+          const { newUrl, parsedDestination } = prepareDestination({
+            appendParamsToQuery: true,
+            destination: result.response.headers.get('x-middleware-rewrite')!,
+            params: _params,
+            query: parsedUrl.query,
+          })
+
+          if (
+            parsedDestination.protocol &&
+            (parsedDestination.port
+              ? `${parsedDestination.hostname}:${parsedDestination.port}`
+              : parsedDestination.hostname) !== req.headers.host
+          ) {
+            return this.proxyRequest(
+              req as NodeNextRequest,
+              res as NodeNextResponse,
+              parsedDestination
+            )
+          }
+
+          if (this.nextConfig.i18n) {
+            const localePathResult = normalizeLocalePath(
+              newUrl,
+              this.nextConfig.i18n.locales
+            )
+            if (localePathResult.detectedLocale) {
+              parsedDestination.query.__nextLocale =
+                localePathResult.detectedLocale
+            }
+          }
+
+          addRequestMeta(req, '_nextRewroteUrl', newUrl)
+          addRequestMeta(req, '_nextDidRewrite', newUrl !== req.url)
+
+          return {
+            finished: false,
+            pathname: newUrl,
+            query: parsedDestination.query,
+          }
+        }
+
+        if (result.response.headers.has('x-middleware-refresh')) {
+          res.statusCode = result.response.status
+          for await (const chunk of result.response.body || ([] as any)) {
+            this.streamResponseChunk(res as NodeNextResponse, chunk)
+          }
+          res.send()
+          return {
+            finished: true,
+          }
+        }
+
+        return {
+          finished: false,
+        }
+      },
+    }
+  }
+
+  private middlewareBetaWarning = execOnce(() => {
+    Log.warn(
+      `using beta Middleware (not covered by semver) - https://nextjs.org/docs/messages/beta-middleware`
+    )
+  })
+
+  protected async runMiddleware(params: {
+    request: BaseNextRequest
+    response: BaseNextResponse
+    parsedUrl: ParsedNextUrl
+    parsed: UrlWithParsedQuery
+    onWarning?: (warning: Error) => void
+  }): Promise<FetchEventResult | null> {
+    this.middlewareBetaWarning()
+
+    // For middleware to "fetch" we must always provide an absolute URL
+    const url = getRequestMeta(params.request, '__NEXT_INIT_URL')!
+    if (!url.startsWith('http')) {
+      throw new Error(
+        'To use middleware you must provide a `hostname` and `port` to the Next.js Server'
+      )
+    }
+
+    const page: { name?: string; params?: { [key: string]: string } } = {}
+    if (await this.hasPage(params.parsedUrl.pathname)) {
+      page.name = params.parsedUrl.pathname
+    } else if (this.dynamicRoutes) {
+      for (const dynamicRoute of this.dynamicRoutes) {
+        const matchParams = dynamicRoute.match(params.parsedUrl.pathname)
+        if (matchParams) {
+          page.name = dynamicRoute.page
+          page.params = matchParams
+          break
+        }
+      }
+    }
+
+    const allHeaders = new Headers()
+    let result: FetchEventResult | null = null
+
+    for (const middleware of this.middleware || []) {
+      if (middleware.match(params.parsedUrl.pathname)) {
+        if (!(await this.hasMiddleware(middleware.page, middleware.ssr))) {
+          console.warn(`The Edge Function for ${middleware.page} was not found`)
+          continue
+        }
+
+        await this.ensureMiddleware(middleware.page, middleware.ssr)
+
+        const middlewareInfo = this.getMiddlewareInfo({
+          dev: this.renderOpts.dev,
+          distDir: this.distDir,
+          page: middleware.page,
+          serverless: this._isLikeServerless,
+        })
+
+        result = await run({
+          name: middlewareInfo.name,
+          paths: middlewareInfo.paths,
+          env: middlewareInfo.env,
+          request: {
+            headers: params.request.headers,
+            method: params.request.method || 'GET',
+            nextConfig: {
+              basePath: this.nextConfig.basePath,
+              i18n: this.nextConfig.i18n,
+              trailingSlash: this.nextConfig.trailingSlash,
+            },
+            url: url,
+            page: page,
+          },
+          useCache: !this.nextConfig.experimental.concurrentFeatures,
+          onWarning: (warning: Error) => {
+            if (params.onWarning) {
+              warning.message += ` "./${middlewareInfo.name}"`
+              params.onWarning(warning)
+            }
+          },
+        })
+
+        for (let [key, value] of result.response.headers) {
+          if (key !== 'x-middleware-next') {
+            allHeaders.append(key, value)
+          }
+        }
+
+        if (!this.renderOpts.dev) {
+          result.waitUntil.catch((error) => {
+            console.error(`Uncaught: middleware waitUntil errored`, error)
+          })
+        }
+
+        if (!result.response.headers.has('x-middleware-next')) {
+          break
+        }
+      }
+    }
+
+    if (!result) {
+      this.render404(params.request, params.response, params.parsed)
+    } else {
+      for (let [key, value] of allHeaders) {
+        result.response.headers.set(key, value)
+      }
+    }
+
+    return result
   }
 }
