+++ packages/next/server/web/sandbox/polyfills.ts
@@ -2,6 +2,7 @@ import { Crypto as WebCrypto } from 'next/dist/compiled/@peculiar/webcrypto'
 import { CryptoKey } from 'next/dist/compiled/@peculiar/webcrypto'
 import { TransformStream } from 'next/dist/compiled/web-streams-polyfill'
 import { v4 as uuid } from 'next/dist/compiled/uuid'
+
 import crypto from 'crypto'
 
 export function atob(b64Encoded: string) {
